Add tempfiles and CSL from URL, remove extra options
I tried making this it's own separate plug-in, but mkdocs will raise exceptions before any files are allowed to be downloaded due to the file exists flag in the config_scheme.

Hopefully this integration looks more okay, it adds the packages `requests` and `validators`.

Now you can either enter a path for bib/csl files, or a URL to download them with tempfile without any added options.