Add new regex to format_citations and quads
The tests look a bit messy now due to `find_cite_keys` returning a tuple, for now this seems to be working.

I've also added the prefix and suffix to pairs in `format_citations`

All inline citations from my testing render as expected, including those with suffixes/prefixes. Affixes are ignored for now, not sure how to pass them correctly to `format_simple|pandoc`, nor understand what the bibtex syntax would be - but this feels either close - or way off :)