Working page numbers in references demo
Got it working afterall, but it's not an ideal solution. This "works" as long as there's only a single citekey and a single reference to page numbers..

Adding this to the inline citation might make more sense.

I presume there must be some sort of pybtex library to parse this automatically from the entire citekey.

Taking a break for now