Mock Zotero API and bibtex entries
Instead of setting up a testing environment in Zotero, I have mocked the
API documented in [Zotero API
docs](https://www.zotero.org/support/dev/web_api/v3/basics) especially
with the pagination feature that splits the bibtex into chunks with a
maximum length that they call 'limit'. In the fixture, the limit is set
to 25 or the default.

Along with the mock API, I needed a function that can generate n bibtex
entries to test paginated and non-paginated results.