Add function to handle Zotero API requests
Zotero API's "Sorting and Pagination"
[docs](https://www.zotero.org/support/dev/web_api/v3/basics) explain
that results are limited to an integer number of results that can
increase to a maximum of 100 results per request. If there are more
results that match the search criteria, a "Link" header will be added
with a `rel=next` link. The `tempfile_from_zotero` function leverages
the response header and continues requesting the `next` url until all
results are returned. To prevent runaway conditions, an arbitrary hard
limit of 999 has been placed.