Sanitize Zotero API URL before request
To prevent the user from adding problematic query params to the Zotero
URL, the `sanitize_zotero_query()` function ensures that the requested
format is bibtex and the limit is set to the maximum permitted by Zotero
to reduce the total number of requests.

The mocked Zotero API needs to represent these latest changes too.